pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # 后端框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",

    # LLM
    "litellm>=1.50.0",

    # 数据库
    "aiosqlite>=0.20.0",

    # 配置
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # MCP
    "mcp>=1.0.0",

    # 工具依赖
    "httpx>=0.27.0",
    "html2text>=2024.2.26",
    "jsonschema>=4.23.0",

    # Skills
    "pyyaml>=6.0.2",

    # 日志
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
